Time: 29 ms (34.82%) | Memory: 38.5 MB (51.78%) - LeetSync

+++ 485-max-consecutive-ones/max-consecutive-ones.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int findMaxConsecutiveOnes(vector<int>& nums) {
-        long count=0;
-        long maximum=0;
-        long n=nums.size();
+        int count=0;
+        int maximum=0;
+        int n=nums.size();
         for(int i=0;i<n;i++){
             if(nums[i]==1){
                 count++;
